Fix: Pass in raw pointer to SetupWizard
Couldn't figure out what was happening with the DatabaseConnection
singleton. It seemed to be destructed after usage in the SetupWizard
because when trying to use it _any_ of the dialogs, it would reinitialize
itself and `pDatabase` pointer would be NULL even though after using
the DatabaseConnection class in SetupWizard and delete the pointer there
AND calling UnsetHandle on the method somehow seemed to influence
the lifecycle of the singleton. Passing in the raw pointer to SetupDatabase
class has resolved the issue.

Might need to come back to this issue one day